+++ doppler/src/panic_handler.rs
@@ -7,13 +7,14 @@ extern "C" {
 macro_rules! nostd_panic_handler {
     () => {
         /// A panic handler for `no_std`.
+        #[cfg(not(feature = "std"))]
         #[cfg(target_os = "solana")]
         #[no_mangle]
         #[panic_handler]
         fn panic_handler(info: &core::panic::PanicInfo<'_>) -> ! {
             if let Some(location) = info.location() {
                 unsafe {
-                    $crate::helpers::sol_panic_(
+                    $crate::panic_handler::sol_panic_(
                         location.file().as_ptr(),
                         location.file().len() as u64,
                         location.line() as u64,
