--- oracle/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "doppler-oracle"
-description = "Oracle for Doppler."
-authors = { workspace = true }
-repository = { workspace = true }
-readme = { workspace = true }
-license-file = { workspace = true }
-edition = { workspace = true }
-version = { workspace = true }
-
-[lib]
-crate-type = ["lib", "cdylib"]
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = [
-  'cfg(target_os, values("solana"))',
-] }
